+++ spec/rubocop/cop/graphql/not_authorized_node_type_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe RuboCop::Cop::GraphQL::NotAuthorizedNodeType, :config do
     end
   end
 
-  context "when type not implements Node interface" do
+  context "when type implements Node interface" do
     context "when type has .authorized? check" do
       specify do
         expect_no_offenses(<<~RUBY, "graphql/types/user_type.rb")
@@ -22,6 +22,24 @@ RSpec.describe RuboCop::Cop::GraphQL::NotAuthorizedNodeType, :config do
 
             def self.authorized?(object, context)
               super && object.owner == context[:viewer]
+            end
+          end
+        RUBY
+      end
+    end
+
+    context "when type has .authorized? check using class << self" do
+      specify do
+        expect_no_offenses(<<~RUBY, "graphql/types/user_type.rb")
+          class UserType < BaseType
+            implements GraphQL::Types::Relay::Node
+
+            field :uuid, ID, null: false
+
+            class << self
+              def authorized?(object, context)
+                super && object.owner == context[:viewer]
+              end
             end
           end
         RUBY